// src/app/api/model/candidateDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CandidateDto { 
    id?: number;
    name?: string;
    votes?: number;
    partyListId?: number;
}